feat #331: rename cards
