+++ components/molecules/bestGames/styled.ts
@@ -9,21 +9,3 @@ export const Container = styled.div`
   max-width: 1000px;
   width: 100%;
 `
-
-export const DesktopCardContainer = styled.div`
-  display: block;
-  width: 100%;
-
-  @media (max-width: 480px) {
-    display: none;
-  }
-`
-
-export const MobileCardContainer = styled.div`
-  display: none;
-  width: 100%;
-
-  @media (max-width: 480px) {
-    display: block;
-  }
-`
